Adds DNI to order details
Adds the customer's DNI (Documento Nacional de Identidad) to the order details displayed in the admin panel and invoice.
This change involves updating the SQL query to retrieve the DNI from the database and displaying it in the relevant views.
Additionally renames billing_barrio to billing_neighborhood and billing_id to dni_cliente to keep the code consistent.